More reorganization
- Moved prjman.lazy to prjman.module
- Removed _LAZY_IMPORTS dictionary
    - Acted the same as sys.modules
- Moved module loading to a standard getter
