+++ src/prjman/metadata/base.py
@@ -5,7 +5,7 @@ import os
 from typing import List, Tuple, Set, Dict
 from pathlib import Path
 import fnmatch
-from prjman.lazy import SINGLETON
+from prjman.module import SINGLETON
 from prjman.metadata.file import ProjectFile
 from prjman.struct.codec import DictCodec, DictObject
 from prjman.utils import get_or_default, input_yn_default
