+++ src/prjman/metadata/files/gitrepo.py
@@ -3,7 +3,7 @@
 
 from typing import Set
 from git import Repo
-from prjman.lazy import SINGLETON
+from prjman.module import SINGLETON
 from prjman.utils import get_default, input_with_default, input_yn_default
 from prjman.struct.codec import DictObject
 from prjman.metadata.file import ProjectFile, ProjectFileCodec, build_file
