+++ src/prjman/metadata/files/osf.py
@@ -2,7 +2,7 @@
 project.
 """
 
-from prjman.lazy import SINGLETON
+from prjman.module import SINGLETON
 from prjman.struct.codec import DictObject
 from prjman.metadata.file import ProjectFile, ProjectFileCodec, build_file
 from prjman.utils import download_and_write
